requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
audio-recorder-streamlit==0.0.8
backoff==2.2.1
bleach==6.0.0
blinker==1.6.2
bullet==2.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chroma-migrate==0.0.3
chromadb==0.4.0
click==8.1.5
click-log==0.4.0
clickhouse-connect==0.6.6
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.12
decorator==5.1.1
deprecation==2.1.0
dnspython==2.4.0
docutils==0.20.1
dotty-dict==1.3.1
duckdb==0.7.1
fastapi==0.99.1
flatbuffers==23.5.26
flupy==1.2.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
gotrue==1.0.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
invoke==1.7.3
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==24.2.0
langchain==0.0.235
langsmith==0.0.7
loguru==0.7.0
lz4==4.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pgvector==0.1.8
Pillow==9.5.0
pinecone-client==2.2.2
pkginfo==1.9.6
plotly==5.15.0
postgrest==0.10.6
posthog==3.0.1
protobuf==4.23.4
psycopg2-binary==2.9.6
pulsar-client==3.2.0
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-semantic-release==7.33.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readme-renderer==40.0
realtime==1.0.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.9.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
storage3==0.5.2
streamlit==1.24.1
StrEnum==0.4.15
supafunc==0.2.2
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.11.8
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
twine==3.8.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.23.1
validators==0.20.0
vecs==0.2.6
watchdog==3.0.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==10.4
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
zstandard==0.21.0
